docs: Move entities directory
Move model.db package to model package.
