+++ src/main/java/com/kurlabo/backend/model/Insta_src.java
@@ -1,4 +1,4 @@
-package com.kurlabo.backend.model.db;
+package com.kurlabo.backend.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
